Move mocks to test utils
