+++ .changeset/strange-flies-retire.md
@@ -1,0 +1,6 @@
+---
+'@keystonejs/keystone': minor
+'@keystonejs/test-utils': patch
+---
+
+Moved adapter mocks into test utils package. No functional changes but if running tests update keystone and test-utils packages.
